Refactor test assertions and add testing extensions
Moved generic assertion helpers to mc.data/extensions/testing_extensions.go for reuse across tests. Updated repo_test.go and av_test.go to use the new helpers, improving consistency and reducing duplication. Added a test for null.Int64 behavior in av_test.go.

+++ mc.data/repos/repo_test.go
@@ -9,8 +9,11 @@ import (
 	"github.com/guregu/null/v6"
 	"github.com/jackc/pgx/v5"
 	"github.com/joho/godotenv"
-	"mc.data/models"
+	ex "mc.data/extensions"
+	m "mc.data/models"
 )
+
+var id null.Int64
 
 func Test_Base_CanGetConnectionAndPing(t *testing.T) {
 	ctx := context.Background()
@@ -23,9 +26,11 @@ func Test_Base_CanGetConnectionAndPing(t *testing.T) {
 }
 
 func Test_TimeSeriesMetaDataRepo_CanInsert(t *testing.T) {
-	ex := models.TimeSeriesMetadata{
+	symbol := "_TEST"
+	
+	testMetaData := m.TimeSeriesMetadata{
 		Information:   null.StringFrom("TEST INFO"),
-		Symbol:        null.StringFrom("_TEST"),
+		Symbol:        null.StringFrom(symbol),
 		LastRefreshed: time.Date(2025, time.October, 31, 0, 0, 0, 0, time.UTC),
 		TimeZone:      null.StringFrom("testtimezone"),
 	}
@@ -33,11 +38,19 @@ func Test_TimeSeriesMetaDataRepo_CanInsert(t *testing.T) {
 	ctx := context.Background()
 	pg := getConnection(t, ctx)
 
-	_, err := pg.InsertNewMetaData(ctx, &ex)
+	idx, err := pg.InsertNewMetaData(ctx, &testMetaData)
+	if err != nil {
+		t.Fatalf("error inserting new meta data: %s", err)
+	}
+
+	id = null.NewInt(idx, true)
+	res, err := pg.GetMetaDataBySymbol(ctx, symbol)
 
 	if err != nil {
-		
+		t.Fatalf("error getting meta data by symbol")
 	}
+
+	ex.AssertAreEqual(t, "information", testMetaData.Information.String, res.Information.Ptr())	
 }
 
 func getConnection(t *testing.T, ctx context.Context) *Postgres {
@@ -56,16 +69,21 @@ func getConnection(t *testing.T, ctx context.Context) *Postgres {
 
 	// on test resolving, this will close connections, even if an error is thrown
 	t.Cleanup(func() {
+		res.deleteTestTimeSeriesData(t, ctx)
 		res.Close()
 	})
 
 	return res
 }
 
-func (pg *Postgres) deleteTestTimeSeriesData(t *testing.T, ctx context.Context, id int64) {
+func (pg *Postgres) deleteTestTimeSeriesData(t *testing.T, ctx context.Context) {
 	t.Helper()
-	sql := `DELETE FROM time_series_data WHERE source_id = @source_id;
-			DELETE FROM time_series_metadata WHERE id = @source_id`
-			args := pgx.NamedArgs{"source_id": id}
-	pg.Execute(ctx, sql, args)
+
+	if id.Ptr() != nil {
+		sql := `DELETE FROM time_series_data WHERE source_id = @source_id;
+				DELETE FROM time_series_metadata WHERE id = @source_id`
+		args := pgx.NamedArgs{"source_id": *id.Ptr()}
+		
+		pg.Execute(ctx, sql, args)
+	}
 }
